refactor: extract shared document validators into src/validators.ts
The workflow mutation is registered in the app but takes and returns the
component's own workflows/steps documents, so the client had to import from
component/schema.ts — pulling the whole defineSchema call in with it.

Move those definitions to src/validators.ts, shared by both sides, and leave
component/schema.ts as just the table definitions. The shared module exports the
boundary shape (string ids, per the previous fix) plus vStepFields/vJournalFields
factories that the schema specializes with real ids for its own tables and for
journal.startSteps, whose args are validated inside the component.
